pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".

[build-system]
requires = [ "poetry_core>=1.0.0" ]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "llm-cdk-app-agent"
version = "0.1.0"
description = "A CDK app for deploying the LLM agent"
authors = [ "Jacob Petterle <jacobpetterle@tai-tutor.team>" ]
readme = "README.md"

  [tool.poetry.dependencies]
  beautifulsoup4 = "4.12.0"
  langchain = "0.0.30"
  openai = "0.28.1"
  pinecone-client = "2.2.4"
  python = "^3.8.1"
  tiktoken = "0.5.1"
  unstructured = "0.10.0"
  watchdog = "^3.0.0"

  [tool.poetry.dev-dependencies]
  projen = "0.76.18"
  pytest = "6.2.1"
